Seperate entrypoint for tempest plugin
Previously python-gnocchi package installed the entry point for the in-tree
Tempest plugin which cause failure while running tempest list-plugins
because Tempest found the entry point but does not have Tempest plugin
code which is subpackaged in python-gnocchi-tests.
Separate gnocchi_tests egg-info is created with the plugin entry point
so that Tempest includes the plugin only when subpackage is installed.

Change-Id: I75f219447765b4d988e18ce1980754235e1b15c1