adjust gnocchi-config-generator in %files common
The location where this entry was added, caused it to
inherit extra group/owner attributes which are not
normally done.  By moving to top of this section it
cause it to have normal permissions and let the rest
of the folders be defined as they were before.

Change-Id: If32578dbabc41d0ef6d6bf1539ce7d9267cb13ab
(cherry picked from commit 17dfa9aca6addfc1b95ee5c52941a75c88ec4c59)