Apply upstream commit 7a3ddc4 as patch
The upstream patch 7a3ddc4 [1] contains
encoding fixes that is required to make
sure everything works with python3.

Since there is no new release containing
this fix and the project is mostly unmaintained
there might not be a new release to rely on.

This is patched the same way in the Ubuntu
packages since there is no release.

[1] https://github.com/gnocchixyz/gnocchi/commit/7a3ddc4ceb64bf648a5f383814632f72b8700a8f

Change-Id: I05773180ee1e5fd0ca3ca373b6c71fa9e8b6ddc2