Ensure python3 encoding issues is fixed
The upstream patch 7a3ddc4 [1] and PR [2]
contains encoding fixes that is required
to make sure everything works with python3.

Since there is no new release containing
these fixes and the project is mostly
unmaintained there might not be a new
release to rely on.

This is patched (with [1]) the same way in
the Ubuntu packages since there is no release.

[1] https://github.com/gnocchixyz/gnocchi/commit/7a3ddc4ceb64bf648a5f383814632f72b8700a8f
[2] https://github.com/gnocchixyz/gnocchi/pull/1064

Change-Id: I05773180ee1e5fd0ca3ca373b6c71fa9e8b6ddc2