Remove upper limit for oslo.policy
Upstream, oslo.policy is capped because of an issue with Ubuntu [1]
which does not affect CentOS. This is breaking deployment in RDO as we
have >= 3.5.0 since Victoria.

This patch is removing this upper limit until it's fixed and removed
upstream.

[1] https://github.com/gnocchixyz/gnocchi/commit/a2d84f5b1b0555ad83b14e931d9bd7065f2f7b6a

Change-Id: I8078cca949fcbaa44564a7e9928cb0c164cb5964
(cherry picked from commit de4443a960c9fa4694847c6bec77ec3a531435e8)