Stop installing policy files
Now gnocchi loads policy.yaml by default, so policy.json is no longer
needed. Also, the current policy.yaml is just empty and we don't have
to explicitly put the file, because gnocchi ignores loading the .yaml
file in case it does not exist (and it result in the same effect as
loading an empty file).

This also removes the sed hacks in setup.cfg, because these have no
effect with recent releases.

Change-Id: I8fa9477a66b9784b3ec54ad15bbcd708f2b4f7fd